// src/types/Synergetic/Finance/iSynVCreditor.ts
type iSynVCreditor = {
  CreditorID: number;
  CreditorIDType: string | null;
  CreditorPayeeName: string;
  CreditorAccountReference: string;
  CreditorPaymentMethod: string;
  CreditorNotifyMethod: string;
  CreditorEFTDetailsFlag: boolean;
  CreditorBankBSB: string;
  CreditorBankAccount: string;
  CreditorOverrideInvDesc: string | null;
  CreditorPaymentTermsDays: number;
  CreditorImmediatePaymentFlag: boolean;
  CreditorSupplierCategory: string;
  CreditorSupplierCategoryDescription: string;
  CreditorSupplierApprovedFlag: boolean;
  CreditorAlertMessage: string;
  CreditorAlertMessageTransFlag: boolean;
  CreditorHoldPaymentFlag: boolean;
  CreditorLastPaymentDate: Date | string | null;
  CreditorLastPaymentAmount: number;
  CreditorLastPaymentChequeNumber: number;
  CreditorCurrentBalance: number;
  CreditorPricesIncludeTaxFlag: boolean;
  CreditorABNFlag: boolean;
  CreditorABN: string;
  CreditorABNDivision: string;
  CreditorGSTRegisteredFlag: boolean;
  CreditorConsignmentFlag: boolean;
  CreditorConsignmentStockPaymentFlag: boolean;
  CreditorMailName: string;
  CreditorMailSalutation: string;
  CreditorNameInternal: string;
  CreditorNameExternal: string;
  CreditorLegalFullName: string;
  CreditorPreferredFormal: string;
  CreditorSurname: string;
  CreditorPhone: string;
  CreditorOccupPhone: string;
  CreditorOccupFax: string;
  CreditorMobilePhone: string;
  CreditorEmail: string;
  CreditorCompanyName: string;
  CreditorMailNamePrimary: string;
  CreditorSuburb: string;
  CreditorState: string;
  CreditorPostCode: string;
  CreditorCountryCode: string;
  CreditorCountryDescription: string;
  CreditorAddress1: string;
  CreditorAddress2: string;
  CreditorAddress3: string;
  CreditorAddressComma: string;
  CreditorAddressFull: string;
  CreditorAddressDPID: string;
  CreditorAddressDPIDBarcode: string;
  CreditorAddressSortPlanNumber: string;
  CreditorAddressPAFVersion: string;
  PurchasesLastYear: number;
  PurchasesYTD: number;
  ActiveFlag: boolean;
}

export default iSynVCreditor;